+++ folio-backend-testing/src/main/java/org/folio/test/extensions/EnableKeycloakDataImport.java
@@ -13,5 +13,9 @@ import org.springframework.test.context.TestPropertySource;
 @Tag("keycloak-import-integration")
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE, ElementType.METHOD})
-@TestPropertySource(properties = {"application.keycloak.enabled=true", "application.keycloak.import.enabled=true"})
+@TestPropertySource(properties = {
+  "application.keycloak.enabled=true",
+  "application.keycloak.import.enabled=true",
+  "application.secure-store.environment=folio",
+  })
 public @interface EnableKeycloakDataImport {}