+++ folio-secret-store/folio-secret-store-starter/src/main/java/org/folio/tools/store/configuration/SecureStoreAutoconfiguration.java
@@ -8,6 +8,7 @@ import org.folio.tools.store.impl.VaultStore;
 import org.folio.tools.store.properties.AwsProperties;
 import org.folio.tools.store.properties.EphemeralProperties;
 import org.folio.tools.store.properties.FsspProperties;
+import org.folio.tools.store.properties.SecureStoreProperties;
 import org.folio.tools.store.properties.VaultProperties;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -16,6 +17,12 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class SecureStoreAutoconfiguration {
+
+  @Bean
+  @ConfigurationProperties(prefix = "application.secret-store")
+  public SecureStoreProperties secureStoreProperties() {
+    return new SecureStoreProperties();
+  }
 
   @Bean
   @ConditionalOnProperty(name = "application.secret-store.type", havingValue = "EPHEMERAL", matchIfMissing = true)